crucible-syntax: Separate interface and implementation of parsing monad (#1133)
Previously, helpers for parsing with the abstract `MonadSyntax` interface
were intermingled with utilities for dealing with the concrete `SyntaxParse`
implementation. This patch splits them up, with the implementation depending
on (and implementing) the interface.
