+++ crucible-syntax/crucible-syntax.cabal
@@ -112,6 +112,7 @@ library
     Lang.Crucible.Syntax.SExpr
     Lang.Crucible.Syntax.Overrides
     Lang.Crucible.Syntax.ExprParse
+    Lang.Crucible.Syntax.Monad
     Lang.Crucible.Syntax.Prog
 
   ghc-options: -Wall -Werror=incomplete-patterns -Werror=missing-methods -Werror=overlapping-patterns
